requirements.txt
peft
accelerate
bitsandBytes
transformers
GPUtil
torch
huggingface_hub
pandas
langchain
faiss-cpu
jq
sentence-transformers
langgraph
langchain-google-genai
python-dotenv
langchain-community